pyproject.toml
[project]
name = "personal-finance-assistant"
version = "0.1.0"
description = "Personal Finance Assistant with AI capabilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-ai",
    "python-dateutil>=2.8.2",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "supabase>=2.10.0",
    "python-dotenv>=1.0.1",
    "google-genai>=0.2.0",
    "pydantic-ai-slim[web]",
    "uvicorn",
    "mlflow",
    "pandas",
    "streamlit",
    "plotly",
    "Pillow",
    "watchdog",
    "asyncpg",
]
